requirements.txt
streamlit
pandas
numpy
pillow
langdetect
geocoder
streamlit
geocoder
langdetect
pillow